requirements.txt
gradio
transformers
sentence-transformers
faiss-cpu
numpy
torch
pypdf
nltk
pytesseract
pillow
